moves Constraint inside the internal package
