+++ src/software/amazon/ionschema/internal/TypeImpl.kt
@@ -2,7 +2,6 @@ package software.amazon.ionschema.internal
 
 import software.amazon.ion.*
 import software.amazon.ion.system.IonSystemBuilder
-import software.amazon.ionschema.Constraint
 import software.amazon.ionschema.Schema
 import software.amazon.ionschema.internal.constraint.ConstraintBase
 import software.amazon.ionschema.internal.constraint.TypeReference
